+++ Test.rs
@@ -1,0 +1,43 @@
+//  TEST.rs
+//    by Lut99
+// 
+//  Created:
+//    30 Jul 2022, 16:56:55
+//  Last edited:
+//    30 Jul 2022, 17:39:57
+//  Auto updated?
+//    Yes
+// 
+//  Description:
+//!   Test 4242
+// 
+
+
+//    by [object Object]
+// 
+//  Created:
+//    30 Jul 2022, 16:56:15
+//  Last edited:
+//    30 Jul 2022, 16:56:18
+//  Auto updated?
+//    Yes
+// 
+//  Description:
+//!   Test 42
+// 
+
+//  UNTITLED 1
+//    by Lut99
+// 
+//  Created:
+//    30 Jul 2022, 12:29:55
+//  Last edited:
+//    30 Jul 2022, 16:48:26
+//  Auto updated?
+//    Yes
+// 
+//  Description:
+//!   Test
+// 
+
+